requirements.txt
numpy
openai
pandas
matplotlib
scikit-learn
tiktoken
transformers
torch
accelerate
torch
torchvision
tensorboard
pyyaml
pre-commit
mypy
black
isort
autoflake
tensorflow
mkl
dtaidistance
seaborn
einops
sentencepiece
omegaconf
pytest
hydra-core
